requirements.txt
# Core dependencies
telethon>=1.24.0
python-dotenv>=0.19.0
pyyaml>=6.0

# Test dependencies
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-cov>=4.0.0

# Development dependencies
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
flake8-docstrings>=1.7.0
pre-commit>=3.3.0